@@ -1,11 +1,15 @@
 package api_operation
 
 import (
+	"encoding/json"
+	"fmt"
 	"strconv"
 
+	"github.com/blevesearch/bleve/v2"
 	"github.com/gin-gonic/gin"
 	"github.com/mrzack99s/cocong/model"
 	"github.com/mrzack99s/cocong/services"
+	"github.com/mrzack99s/cocong/vars"
 )
 
 func (ctl *controller) loginLogQuery(c *gin.Context) {
@@ -65,8 +69,6 @@ func (ctl *controller) networkLogQuery(c *gin.Context) {
 	search := c.Query("search")
 	offsetStr := c.Query("offset")
 	limitStr := c.Query("limit")
-	or := c.Query("or")
-
 	offset, e := strconv.Atoi(offsetStr)
 	if e != nil {
 		c.String(400, "offset is not correct, allow only integer")
@@ -79,35 +81,55 @@ func (ctl *controller) networkLogQuery(c *gin.Context) {
 		return
 	}
 
-	response := []model.NetworkLog{}
-	count, err := services.DBQuery(&response, offset, limit, search, or == "true", false)
+	query := bleve.NewMatchQuery(search)
+	searchRequest := bleve.NewSearchRequest(query)
+
+	// กำหนด From และ Size สำหรับ Pagination
+	searchRequest.From = offset
+	searchRequest.Size = limit
+
+	// ค้นหาข้อมูล
+	searchResult, err := vars.NetLogDatabase.Search(searchRequest)
 	if err != nil {
-		c.String(500, err.Error())
+		c.String(500, fmt.Sprintf("Search error: %v", err))
 		return
 	}
 
-	c.JSON(200, gin.H{
-		"Count": count,
-		"Data":  response,
-	})
+	result_doc := []model.NetworkLog{}
+	for _, hit := range searchResult.Hits {
 
-}
+		docBytes, err := vars.NetLogDatabase.GetInternal([]byte(hit.ID))
+		if err != nil {
+			c.String(500, "Error fetching document")
+			return
+		}
 
-func (ctl *controller) networkLogDump(c *gin.Context) {
+		if docBytes == nil {
+			c.String(500, "Document not found")
+			return
+		}
 
-	search := c.Query("search")
-	or := c.Query("or")
+		// แปลงข้อมูลจาก JSON เป็น Document struct
+		var doc model.NetworkLog
+		err = json.Unmarshal(docBytes, &doc)
+		if err != nil {
+			c.String(500, "Error unmarshalling document")
+			return
+		}
 
-	response := []model.NetworkLog{}
-	count, err := services.DBQuery(&response, 0, 0, search, or == "true", true)
-	if err != nil {
-		c.String(500, err.Error())
-		return
+		result_doc = append(result_doc, doc)
 	}
 
+	// response := []model.NetworkLog{}
+	// count, err := services.DBQueryCustomDB(vars.NetLogDatabaseRO, &response, offset, limit, search, or == "true", false)
+	// if err != nil {
+	// 	c.String(500, err.Error())
+	// 	return
+	// }
+
 	c.JSON(200, gin.H{
-		"Count": count,
-		"Data":  response,
+		"Count": searchResult.Total,
+		"Data":  result_doc,
 	})
 
 }
