@@ -40,29 +40,40 @@ func tokenMiddleware(c *gin.Context) {
 	c.Next()
 }
 
+func tokenMiddlewareSessionOnly(c *gin.Context) {
+	tokenString := c.Request.Header.Get("api-token")
+
+	_, ok := vars.AdminSession.Get(tokenString)
+	if !ok {
+		c.JSON(http.StatusUnauthorized, "not authorised")
+		c.Abort()
+		return
+	}
+
+	c.Next()
+}
+
 func (ctl *controller) register() {
 
 	ctl.router.POST("/op/login", ctl.login)
 	ctl.router.DELETE("/op/logout", ctl.logout)
 	ctl.router.POST("/op/refresh-token", ctl.refreshToken)
-	ctl.router.GET("/op/me", tokenMiddleware, ctl.me)
-	ctl.router.POST("/op/change-password", tokenMiddleware, ctl.adminChangePassword)
+	ctl.router.GET("/op/me", tokenMiddlewareSessionOnly, ctl.me)
+	ctl.router.POST("/op/change-password", tokenMiddlewareSessionOnly, ctl.adminChangePassword)
 
 	session := ctl.router.Group("/op/session")
-	session.Use(tokenMiddleware)
-
-	session.GET("/query", ctl.sessionQuery)
-	session.PATCH("/kick", ctl.sessionKick)
+	session.GET("/query", tokenMiddleware, ctl.sessionQuery)
+	session.PATCH("/kick", tokenMiddlewareSessionOnly, ctl.sessionKick)
 
 	bw := ctl.router.Group("/op/bandwidth")
-	bw.Use(tokenMiddleware)
+	bw.Use(tokenMiddlewareSessionOnly)
 
 	bw.GET("/query", ctl.bandwidthQuery)
 	bw.POST("/create", ctl.bandwidthCreate)
 	bw.DELETE("/delete", ctl.bandwidthDelete)
 
 	directory := ctl.router.Group("/op/directory")
-	directory.Use(tokenMiddleware)
+	directory.Use(tokenMiddlewareSessionOnly)
 
 	directory.GET("/query", ctl.directoryQuery)
 	directory.POST("/create", ctl.directoryCreate)
@@ -70,7 +81,7 @@ func (ctl *controller) register() {
 	directory.DELETE("/delete", ctl.directoryDelete)
 
 	user := ctl.router.Group("/op/user")
-	user.Use(tokenMiddleware)
+	user.Use(tokenMiddlewareSessionOnly)
 
 	user.GET("/query", ctl.userQuery)
 	user.POST("/create", ctl.userCreate)
@@ -79,7 +90,7 @@ func (ctl *controller) register() {
 	user.PATCH("/password-reset", ctl.userPasswordReset)
 
 	administrator := ctl.router.Group("/op/administrator")
-	administrator.Use(tokenMiddleware)
+	administrator.Use(tokenMiddlewareSessionOnly)
 
 	administrator.GET("/query", ctl.administratorQuery)
 	administrator.POST("/create", ctl.administratorCreate)
@@ -88,12 +99,20 @@ func (ctl *controller) register() {
 	administrator.PATCH("/password-reset", ctl.administratorPasswordReset)
 
 	logs := ctl.router.Group("/op/log")
-	logs.Use(tokenMiddleware)
+	logs.Use(tokenMiddlewareSessionOnly)
 
 	logs.GET("/login", ctl.loginLogQuery)
 	logs.GET("/logout", ctl.logoutLogQuery)
 	logs.GET("/net", ctl.networkLogQuery)
 	logs.GET("/login-dump", ctl.loginLogDump)
 	logs.GET("/logout-dump", ctl.logoutLogDump)
-	logs.GET("/net-dump", ctl.networkLogDump)
+
+	sys := ctl.router.Group("/op/system")
+	sys.Use(tokenMiddlewareSessionOnly)
+
+	sys.GET("/config", ctl.getConfig)
+	sys.PUT("/config", ctl.writeConfig)
+
+	sys.PATCH("/service/core/restart", ctl.restartService)
+	sys.PATCH("/service/dns/restart", ctl.restartDNSServer)
 }
