@@ -5,7 +5,7 @@ import (
 	"time"
 
 	"github.com/gin-gonic/gin"
-	"github.com/mrzack99s/cocong/model/inmemory_model"
+	"github.com/mrzack99s/cocong/session"
 	"github.com/mrzack99s/cocong/vars"
 )
 
@@ -13,19 +13,21 @@ func (ctl *controller) getStatus(c *gin.Context) {
 
 	clientIp := c.ClientIP()
 
-	mSession := inmemory_model.Session{}
-	err := vars.InMemoryDatabase.Where("ip_address = ?", clientIp).First(&mSession).Error
+	mSession, err := session.Instance.GetByIP(clientIp)
 	if err != nil {
 		c.Redirect(http.StatusFound, "/login")
 		return
 	}
 
 	if mSession.BandwidthID != nil {
-		vars.Database.Where("id = ?", *mSession.BandwidthID).First(&mSession.Bandwidth)
+		vars.Database.Where("id = ?", mSession.BandwidthID).First(&mSession.Bandwidth)
 	}
 
-	var concurrent int64
-	vars.InMemoryDatabase.Model(&inmemory_model.Session{}).Select("count(id)").Where("user = ?", mSession.User).Scan(&concurrent)
+	concurrent := 0
+	listUsernameSession, err := session.Instance.GetByUsername(mSession.User)
+	if err == nil {
+		concurrent = len(listUsernameSession)
+	}
 
 	c.HTML(200, "status.html", map[string]any{
 		"ChangePasswordPageEndpoint": ChangePasswordPageEndpoint,
@@ -36,6 +38,7 @@ func (ctl *controller) getStatus(c *gin.Context) {
 		"Concurrent":                 concurrent,
 		"Session":                    mSession,
 		"SessionLastSeen":            mSession.LastSeen.Format(time.RFC822),
+		"RedirectURL":                vars.Config.RedirectURL,
 	})
 
 }
