@@ -2,12 +2,13 @@ package setup
 
 import (
 	"fmt"
+	"os"
 	"time"
 
+	"github.com/blevesearch/bleve/v2"
 	"github.com/mrzack99s/cocong/constants"
 	"github.com/mrzack99s/cocong/drivers/sqlite"
 	"github.com/mrzack99s/cocong/model"
-	"github.com/mrzack99s/cocong/model/inmemory_model"
 	"github.com/mrzack99s/cocong/utils"
 	"github.com/mrzack99s/cocong/vars"
 	"gorm.io/gorm"
@@ -18,13 +19,29 @@ func Database() {
 	var err error
 
 	dbname := "cocong.db"
+	netlog_dbname := "cocong-netlog.dblog"
+	authlog_dbname := "cocong-authlog.dblog"
 	if !vars.SYS_DEBUG {
-		dbname = constants.APP_DIR + "/cocong.db"
+		if vars.Config.CustomDBDir != "" {
+			dbname = vars.Config.CustomDBDir + "/cocong.db"
+			netlog_dbname = vars.Config.CustomDBDir + "/" + netlog_dbname
+			authlog_dbname = vars.Config.CustomDBDir + "/" + authlog_dbname
+		} else {
+			dbname = constants.APP_DIR + "/cocong.db"
+			netlog_dbname = constants.APP_DIR + "/" + netlog_dbname
+			authlog_dbname = constants.APP_DIR + "/" + authlog_dbname
+		}
+
+	}
+
+	dsnTail := "?cache=shared&_busy_timeout=5000&_journal_mode=WAL&_synchronous=OFF&_temp_store=MEMORY&_cache_size=-1000000"
+	if vars.Config.DBCacheSize > 0 {
+		dsnTail = fmt.Sprintf("?cache=shared&_busy_timeout=5000&_journal_mode=WAL&_synchronous=OFF&_temp_store=MEMORY&_cache_size=-%d", vars.Config.DBCacheSize)
 	}
 
 	if !vars.SYS_DEBUG {
-		key := "Brey@crazu!risweqI==4!boyu?oqE*As+Spenu7i#OtaYif!i$RokE=h*PlDr@p"
-		dbnameWithDSN := dbname + fmt.Sprintf("?_pragma_key=%s&_pragma_cipher_page_size=4096", key)
+
+		dbnameWithDSN := dbname + dsnTail
 		vars.Database, err = gorm.Open(sqlite.Open(dbnameWithDSN), &gorm.Config{
 			NowFunc: func() time.Time {
 				return time.Now().In(vars.TZ)
@@ -36,6 +53,37 @@ func Database() {
 		if err != nil {
 			panic(err)
 		}
+
+		if _, err := os.Stat(netlog_dbname); os.IsNotExist(err) {
+			vars.NetLogDatabase, err = bleve.NewUsing(netlog_dbname, bleve.NewIndexMapping(), "scorch", "scorch", nil)
+			if err != nil {
+				panic(err)
+			}
+
+		} else {
+
+			vars.NetLogDatabase, err = bleve.Open(netlog_dbname)
+			if err != nil {
+				panic(err)
+			}
+
+		}
+
+		if _, err := os.Stat(authlog_dbname); os.IsNotExist(err) {
+			vars.AuthLogDatabase, err = bleve.NewUsing(authlog_dbname, bleve.NewIndexMapping(), "scorch", "scorch", nil)
+			if err != nil {
+				panic(err)
+			}
+
+		} else {
+
+			vars.AuthLogDatabase, err = bleve.Open(authlog_dbname)
+			if err != nil {
+				panic(err)
+			}
+
+		}
+
 	} else {
 		vars.Database, err = gorm.Open(sqlite.Open(dbname), &gorm.Config{
 			NamingStrategy: schema.NamingStrategy{
@@ -47,18 +95,13 @@ func Database() {
 		}
 	}
 
-	inmemoryDBName := "file::memory:?cache=shared"
-	if vars.SYS_DEBUG {
-		inmemoryDBName = "inmemory.db"
-	}
-	vars.InMemoryDatabase, err = gorm.Open(sqlite.Open(inmemoryDBName), &gorm.Config{
-		NamingStrategy: schema.NamingStrategy{
-			SingularTable: true,
-		},
-	})
+	sqlDB, err := vars.Database.DB()
 	if err != nil {
 		panic(err)
 	}
+
+	sqlDB.SetMaxOpenConns(0) // Maximum number of open connections
+	sqlDB.SetMaxIdleConns(0) // Maximum number of idle connections
 
 	dbmigrate()
 	dbinit()
@@ -73,15 +116,9 @@ func dbmigrate() {
 		model.AdministratorLoginLog{},
 		model.Directory{},
 		model.User{},
-		model.NetworkLog{},
-	}
-
-	inmemoryModels := []any{
-		inmemory_model.Session{},
 	}
 
 	vars.Database.AutoMigrate(models...)
-	vars.InMemoryDatabase.AutoMigrate(inmemoryModels...)
 }
 
 func dbinit() {
